export error fix
