+++ packages/db/src/index.ts
@@ -1,17 +1,2 @@
-import { PrismaClient } from "@prisma/client";
-
-const prismaClientSingleton = () => {
-  return new PrismaClient();
-};
-type PrismaClientSingleton = ReturnType<typeof prismaClientSingleton>;
-
-// eslint-disable-next-line
-const globalForPrisma = globalThis as unknown as {
-  prisma: PrismaClientSingleton | undefined;
-};
-
-const prisma = globalForPrisma.prisma ?? prismaClientSingleton();
-
-export { prisma };
-
-if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
+export { prisma } from "./client";
+export * from "@prisma/client";
