+++ packages/schema/package.json
@@ -20,6 +20,6 @@
     "typescript": "^5.3.3"
   },
   "exports": {
-    "./*": "./src/*.ts"
+    "./*": "./dist/*.js"
   }
 }
